REL-1501: Changed way scale works to avoid floating point issues (via rationals) and to not generate the 1/1x scale in the zoom out menu.

+++ bundle/jsky.app.ot/src/main/java/jsky/image/gui/ImageDisplayMenuBar.java
@@ -37,7 +37,7 @@ public class ImageDisplayMenuBar extends JMenuBar {
     private static final I18N _I18N = I18N.getInstance(ImageDisplayMenuBar.class);
 
     /** Maximum scale (zoom) factor for menu. **/
-    public static final float MAX_SCALE = 20.0F;
+    public static final int   MAX_SCALE = 20;
 
     /** Minimum scale (zoom) factor for menu. **/
     public static final float MIN_SCALE = 1.0F / MAX_SCALE;
